Code for first task of ninth day finished

+++ Day05Task1Result/Program.cs
@@ -9,7 +9,7 @@ namespace Day05Task1Result
     {
         static void Main(string[] args)
         {
-            IList<int> result = Solution.Run(Data.Program, 1);
+            IList<long> result = Solution.Run(Data.Program, 1);
             if (result.Take(result.Count - 1).Any(r => r != 0))
             {
                 throw new Exception("Not all tests passed.");
